Unable to listen to callbacks when loading JSON file Issue before the fix: when loading a json file that has been marked as passthrough in ember-cli-mirage, the AJAX success and error handler doesn't get fired. Normal api calls that is not configured as passthrough works fine.
What changed:
We upgraded ember and all related add ons. There seems to be an issue with the pretender and jQuery.
https://github.com/pretenderjs/pretender/issues/85

ember-cli-mirage has listed jQuery version 1.11.3 as its dependency. so when we upgraded the addons, jQuery got upgraded to 1.11.3.

But pretender which is also a dependency for ember-cli-mirage, is having issues with jQuery versions 1.x

How things work now:
upgraded jquery to 2.1.4. We should be able to load mock data from json file and be able to attach success and failure callbacks

Issue: ASOC-12642