Threats UI authentication using SA
Issue Before:
Only could authenticate using the hard-coded users. Existing logins in
the SA UX did not work.

What changed:
Authenticate user using SA UX. Handle error conditions on the
account status and display it to the user. Make the roles,
permissions, and user's display name available to the new
stack. Also, renamed the session cookie's name to
"SASESSIONID" so it does not clash with SA.

How things work now:
Login credentials are forward to SA UX, where the authentication is
performed. On success, user's display name and permissions are
gathered and then an authentication object containing them are returned
to SA-UI. On failure, the error condition is attached to the response header
and error message is in the response body are returned to SA-UI.

Issue: ASOC-13656