Visual Issues When Loading Style Guide
Issue before the fix:
A few of the pages in the style guide pages were not updating properly when transitioning from the previous route. This was because the code property was on the wrong object. It was also possible to get an invalid theme or locale based on what was in local storage.

What changed:
componen-lib.json was updated to fix the transition error in the guide, and the application view was updated to verify the selected theme and locale in localStorage is a valid option.

How things work now:
An invalid theme or locale will fall back to the defaultLocale, and the style guide pages all transition properly.

Issue: ASOC-16985