Add Friendly Username
Issue before the fix:
The system only had the ability to show the user’s username that is stored in the session. That name might not be very readable.

What changed:
The ability to customize your displayed username was added. Also, there were minor style updates to the header and footer per UX review with Brian.

How things work now:
In the User Preferences Panel, you can elect to set a friendly/readable username. In the case where that is set, it will be displayed. When it is not set, the app will default to the username stored in the user’s session.

Issue: ASOC-19171