Add progress-bar mode to badge-score component
Issue before the fix:
The component only presented a single way to visualize the score.

What changed:
Inline mode was added to component-lib.

How things work now:
When using the content-badge-score component if isInline is set to true, the visualization changes to display the score on the left and a progress bar on its right.

Issue: ASOC-20617