Display number of events
Issue before fix:
There was no tie in to the model from the template for the count of events.

What changed:
There is now a tie in to the model from the template for the count of events. Also for both alerts and events, if no value is found, then "-" is put in place.

How it works now:
If there is an event count, display the count, else display ‘-‘

Issue: ASOC-21498