OAuth Login with Ember Simple Auth (#196)
Issue before the fix:
The authentication mechanism we were using did not work with OAuth.

What changed:
Used additional features in the Simple Auth library to support OAuth in production

How things work now:
On a production build, when configured with Security Server and API Gateway,
the login credentials will be sent to the Security Server.  If successful, an OAuth
token will be returned else failure message.  The JWT token will be saved and
passed to subsequent requests to NW-UI services.

On a development build aka 'ember serve...', the login credentials will be sent to
'/api/user/login' on the proxied server.

Issues: ASOC-20758