ASOC-26538 and ASOC-26537 Carousel IE-11 fixes (#597)
Issue:
IE-11 does not properly display cards on the carousel. The cards float off the page and do not contain the correct size.
IE-11 does not properly display score in the Incident Detail.

How it works now:
Cards are now properly displayed (in table layout style) and are proper height. Also addressed height as it pertains to Incident Detail view.