ASOC-27186-show-dash-when-ipsrc-and-ipdst-not-available (#611)
* ASOC-27186-show-dash-when-ipsrc-and-ipdst-not-available

Issue:
There needs to be a default display for ip-source and ip-destination information. When data is not available the view should display '--'. Additionally, a tooltip should be provided to provide the user with more information about information displayed.

How it works:
Added default display of '--' to rsa-content-ip-connections component that is responsible for displaying ip-source and ip-destination information.
Added a tooltip to provide information about the content.

JIRA:
ASOC-27186

* ASOC-27186

Issue:
Changes to component needed to be more versatile to support existing functionality while also allowing to display a default state.

How it works:
Updated component to support a boolean flag that determines whether default place holder text should be shown.

* ASOC-27186 Addressed PR comments: Refactor showArrow computed property.

* ASOC-27186 Changes made based on UX feedback. Changed tooltip to title. Changed double dash to single dash and updated CSS.