ASOC-27495 Handle back click by preserving event overview panel state. (#705)
* ASOC-27495 Handle back click by preserving event overview panel state.
Issue:
User does not navigate back to Event Overview panel when they have transitioned away from that view (eg. to Investigation page)

Changes:
To re-create the page view the state of the view is saved into local storage for the purpose of re-creating it when user transitions back into incident details view (specificially details/route).

Note:
First approach was done with queryParams. This was a desirable because it delegated to the Ember framework to handle view management as it pertains to navigation history. However, a couple of issues prevented this implementation:

1. Known queryParams bug https://github.com/emberjs/ember.js/issues/10577 prevented usage of replace flag which resulted in saving all of navigation history. This was not desirable because it led to new issues where navigation should not be stored in history.

2. Due to view manipulation from details/route and incident/route it became necessary to use an event-bus to handle changes in queryParams (reseting them). This also lead to convoluted logic.

As such, an alternative approach was taken to use an existing service (respond-mode) that works with local storage to save event overview state.

* Addressed PR comments