ASOC-28688: EmberJS - Live - Add details panel
Issue before the fix:
We have a list of search results, but we only show six columns of data, whereas the
individual resources have far more information associated with them, including the
dependencies, the size, and other metadata.

What changed:
Added a new details panel that shows the resource's details when the resource row
is clicked by the user in the data table.

Issue: ASOC-28688