ASOC-28677: Create Wizard Component (#765)
Issue before the fix:
Upcoming features of Live Search involve a wizard-driven step-by-step workflow for
deploying Live Content resources to an environment. This requires a generic, wizard
component for configuring one or more steps for accomplishing a task.

What changed:
A new Wizard component was added. The Wizard acts as a wrapper for a series of
step components that can be added to the Wizard to create a workflow of sequential
steps that must be completed by the user. Any component can become a Wizard step
by extending the Wizard-Step mixin, and following the proper component declaration
process.

The Wizard displays three discrete regions: progress bar, content, button bar. The
progress bar displays a summary of all the steps and indicates which step is the
user's current step in the workflow. The content region displays the content for the
step. This is the content of the component(s) that have extended the Wizard Step mixin
and which was configured in the Wizard component declaration. The button bar region
provides a set of buttons for navigating forward and backwards in the wizard workflow,
as well as finishing the workflow altogether (after the last step).

The Wizard is built to allow for invalid steps. If a step is invalid, the user cannot
progress to any subsequent step until the problem has been resolved.

Issue: ASOC-28677