RSA Explorer Component (#1138)
Issue before change:
* The Respond pages are currently showing a lot of overlap. We started
with the Incidents list page where we have a table showing all of (or a
subset of) the incidents We had a filter panel with a set of controls
(e.g., date picker, reset) used to filter the result set of incidents.
We had a toolbar across the top with bulk actions for selected
incidents.
* We also have newer pages that are arriving in the short or near term
(e.g., alerts and remediation-tasks) that need all of the same
features, with the addition of a preview panel/inspector (which may
actually be useful if added to the incidents page).
* We don’t want to copy and paste all of this code for every single new
page. We want to reuse as much as possible.

What changed:
* Created a component called `rss-explorer`. The `rss-explorer`
component is a high-level component that has built into it all of the
aforementioned concepts (toolbar, filter panel, a data table with
sorting in column headers and select checkboxes, and a
preview/inspector pane.
* The `rss-explorer` component wraps all of this up, and also builds in
the redux action creators and computed properties, so that the
consuming / implementing parent component does not need to implement
all of the same stuff.
* The `rss-explorer` component also is constructed using the
`contextual component` pattern and bakes in child components, while at
the same time yielding them up to the blocks of the consumer, which can
change attributes and control content.

Notes:
* One new concept introduced here to help DRY up constantly repeating
patterns is a `namespace` that gets supplied to the `rss-explorer` on
declaration. The `namespace` is then used by the explorer to
dynamically resolve computed properties (stateToComputed) and action
creators (dispatchToActions)
* Unit/Integration tests will follow in another PR
* The first implementation of `rss-explorer` will follow shortly as
well, namely the Remediation Tasks page.

Issues: ASOC-33573