Category Picker: Filter by parent property (#1171)
* Category Picker: Filter by parent property

Issue before fix:
* Once more with feeling, as they say. We originally had been trying to provide
a category filter on the Incidents filter panel that would filter the Incidents
by leaf categories. When it turned out that leaf categories were not unique, we
looked at filtering by both the parent and name properties on the categories.
This did not work in the microservice and a bug was filed, but we added the
client side changes in preparation for a service fix. Then it came about that
the requirement here to filter by leaf categories is unnecessary and we only
need to filter by root categories (the parent). This does work already without
any service changes.
* Some minor variable name changes and RSVP use syntax improvements were
requested in a previous PR

What changed:
* Modified the code back to using a single property filter for category
(namely category.parent)
* Fixed a copy-pasted variable name and adjusted rsvp resolve call to be
simpler

Issue: ASOC-34209

* Adjusted filters test to account for only parent cats