Handle for password change requirement during password migration (#1207)
* Handle for password change requirement during password migration

Issue before the fix:
If the password was hashed using the old standard, the password is re-hashed
using the new standard.  During the migration, if the password fails to meet the
password policy an error is thrown.  This error was not captured by the UI to show
the change password screen.

How things work now:
Handle the exception to show the change password screen.

ASOC-33619

* lint: fix quotes