Retrofit Incidents to Explorer with Deletions (#1214)
* Retrofit Incidents to Explorer with Deletions

Issue before fix:
* The incidents list page was originally built as a bespoke page that
was highly customized to the Incidents domain. Last sprint I introduced
the `Explorer` component, which abstracted out all of the common
elements between Incidents and the new Remediation Tasks page, since
the Remediation Tasks page had (from a design and interaction
perspective) an enormous amount of overlap with Incidents (specifically
the same layout with filters, table listing, and toolbar). Moreover,
the same basic operations/actions existed between the pages, namely the
fetching of items, filtering of items, selecting and sorting items, and
the ability to focus on one specific item. Also, the `Explorer`
component introduced a standard "interface" of redux reducers and state
which can be reused across all pages that implement the Explorer
component. Consequently, we need to retrofit Incidents so that it, too,
leverages the Explorer code-base, which in turn allows us to delete all
of the bespoke Incidents code.
* The `Explorer` component also introduced the concept of the
`Inspector`, which displays individual item properties/details, and (as
introduced in `Remediation Tasks`), allows for actions such as editing
item properties. Originally, the Incidents list page did not have an
Inspector, but instead allowed for editing priority, status, and
assignee using an `action-drawer`, that slid out from the right on
hover of the table's item row. Because `Explorer` introudces the
Inspector as a place for making individual item changes, it makes sense
to remove the old `action-drawer` and replace with the standard
`Explorer Inspector`.
* Incidents Sorting: Originally the sorting of Incidents was done in
the Incidents Toolbar using a dropdown menu that listed all of the
possible sorting options. UX feedback indicated that all sorting should
be done via the column headers, which is currently how it works in
`Remediation Tasks`, and which is also a built-in part of the
`Explorer` component.
* Select All: Originally there was no `select-all` toggle checkbox in
the header row cell for the select row column. We need to be able to
select all known rows (maximum of 1000) that are on the page. The
`select-all` checkbox also provides toggling so that we can remove the
Deselect All button introduced in a previous sprint. `Explorer`
provides this as part of it's standard interface.
* Deletions: There was no mechanism by which Incidents could be
deleted, and this should be a standard part of the Explorer interface
(at least in terms of the reducer interface)
* Confirmation Dialogs: `Explorer` needs to provide a mechanism for
standard bulk update and delete confirmation dialogs which can be
reused across all consumers of the `Explorer` component.
* The Date Range selection dropdown had a search box that didn't work
and could be removed.

What changed:
* A lot
* This PR is huge for a number of reasons. First, we see a lot of
code/files being deleted because they are now unused. The entire
`rsa-respond-incidents` component and child components disappear, for
example, along with other components and utilities that are now
obsolete (e.g., the `update-action-button`, sorting options file, test
files, etc). Second, a very large number of files needed to be touched
or created as they were reorganized to meet `Explorer`'s expected
component and reducer/creator/state interface. Third, the sizable
changes introduced a fair number of bugs that were individually
addressed as part of the dev and testing process, meaning more file
changes as needed. And finally, because a large number of files were
being modified, I used this as an opportunity to comb through the code
for variables that needed renaming, methods that needed refactoring.
* Many of the changes are minor (such as the renaming of methods or the
copying of files wholesale from one location to another)
* A breakdown of the changes includes:
* Migrating all of incidents-related action creators to the
`respond/actions/creators/incidents-creators.js` file and removing all
references to the old `data-creators.js` and `ui-state-creators.js`,
since it became clear that these files were going to grow too large and
hold too much disparate content.
* Converting the incidents reducer to use the common explorer reducer
functions, which is also now shared with Remediation Tasks (and in the
future, Alerts)
* Renaming the incident creator functions to match the `Explorer`
interface
* Adding new Action Types for Incidents to match Explorer's
functionality (e.g., Toggle Select All)
* Adding the `rsa-incident\overview` to the Incidents inspector so that
we can now see the incident overview details from the listing page
* Making the `rsa-incident\overview` component a "dumb" component by
removing `stateToComputed`. Instead, all state will be fed to the
component
* Updating the `rsa-incident\overview` component so that it can have
'edit-butttons' for changing priority, assignee, and status
* Deleting the entire `rsa-incident\update-action-button`, since it was
a bit of a mess (tried to do too much), and replacing with a simple
Power Select button implementation pattern that was first implemented
in Remediation Tasks's inspector panel.
* Forklifting the `Incident Filters` to the `Explorer Filters` panel
* Implementing an `Explorer`-based `deleteItem` interface (which covers
both single item deletions and bulk deletions)
* Implementing confirmation modal/dialogs for Bulk Update and Delete,
and embedding them into the Explorer component for reuse
* Re-implementing the incidents toolbar with the same Power Select
button implementation pattern used in the Inspector single-item updates
* Adding Delete button to Incidents
* i18n mapping for en and ja
* CSS additions/modifications for all changes listed above
* Removed search box on Explorer Date Range power select

Notes:
* The Remediation Tasks page does not have a delete bulk (or delete
individual) option yet. Forthcoming
* Many tests were deleted, and while many of them are obsolete, some
need to be redone with Explorer in mind. Forthcoming.
* There is still no delete button in the Inspector, although we expect
to see that incorporated for individual item deletion

Issues: ASOC-34758, ASOC-31593, ASOC-33372, ASOC-34262, ASOC-35004
(partial)

* Fixed i18n mismatch

* Reworking of Explorer after input from GB

* Support Delete on Remediation Tasks page

* Fixed tests

Created a simple stub component that can stand in as the default
component type for Explorer yielded contextual components.