Component: editable-field (#1237)
* Component: editable-field

Issue before fix:
* We have some user stories that involve free text editing of text
properties on items (e.g., remediation task assignee or title/name).

What changed:
* Added a new `respond-common/editable-field` component.
* The display is intended to be similar to Jira’s nice implementation,
where they show the text on the page as is, but when you hover over the
text, you see that the text has a border and an edit button. Then, when
you click the text/button, you enter edit mode, which allows you to
modify the text, and then click either a confirm or cancel button.

Notes:
* If the user enters edit mode but doesn’t change anything and then
focuses out (blurs) from the input, the field exits edit mode.
* When the user clicks the confirm button, an action is sent
(‘fieldChanged’), which includes the following arguments: (1) the new
value, (2) the original value, (3) a callback function that will revert
the field back. The callback is provided as a convenience in case a
server call is being made with the changed info, but the call fails.
The developer can then simply execute the revert callback function, and
the state is reverted to the original starting point.

Issues: ASOC-35018

* Minor fixes

* Fixed eslint, which meant correcting the template, in which I had
removed space between template elements to prevent inline elements from
having the default spacing, and instead went with a css solution
(involving font-size 0 on the parent element).
* Also took GB’s advice and reverted to a function onFieldChange on the
component, rather than a sendAction() for informing consumers of the
change.