Respond Alerts: Filters (#1264)
* Respond Alerts: Filters

Issue before fix:
* We only saw the OOTB Explorer filter in the Alerts filters panel
(namely, the created/received time range filter).
* The alert source ECAT was appearing in the grid, but we want to show
that as “Endpoint”
* The alert overview in the inspector was not showing the alert name,
and the time shown was not the correct one.

What changed:
* Added a number of filter controls for Alerts. Specifically: an Alert
Type filter, Alert Source filter, a Severity range slider, and a Part
of Incident filter.
* For alert type and alert source, I added a set of hardcoded values
that are returned as a resolved promise in the dictionary api calls
getAllAlertTypes() and getAllAlertSources(). It’s expected that these
will be replaced with server calls eventually (or replaced with facet
calls). Also added action types and reducers for these.
* Updated the fetchItemsStreamBatchRetrieved reducer function to check
if the data returned in the payload is null, and if so, set to empty
array. This won’t be necessary once the service is set to always return
an empty array when there are no results.
* Added some handling in the explorer-build-query.js for filters that
include the type === ‘range’. Some of the original custom date range
filtering will eventually be converted to use this same filter
signature.
* Updated column widths to better manage grid space

Notes:
* The range slider does not currently work because the micro service
does not handle ranges that are not date ranges. Added backend dev task
ASOC-36102 to get that fixed (or to have the service offer an
alternative solution)
* The count at the table footer does not show properly, since there is
a missing service call for getting the alert count separately from the
results (which is how incidents and remediation tasks work). There is a
backend dev task to fix this (ASOC-36010)
* Currently the stream limit of 1000 is not respected by the service.
Added backend dev task ASOC-36022 to get that fixed.

Issues: ASOC-32372

* Fixed label on source (was showing 'Type')

* Fixed missing i18n key/value in ja/trans-data.js