Remediation fixes (#1273)
* Respond - Remediation: Minor Fixes

Issue before fix:
* If an editable-field is empty and there is no placeholder, the
display of the field on hover is all wonky.
* The remediation title editable field needs a placeholder
* Previous changes to the initial state for incidents/tasks/alerts
caused persisted filters to be lost.

What changed:
* Adjusted CSS for the editable-field component so that it has a
min/max width/height that will look nice on hover even when the field
is empty string or null
* Added a placeholder for remediation title in case it is missing
altogether:  (empty)
* Updated reducer handling so that initial state from storage does not
get overridden by setting of default date range filter.

Issues: N/A

* Updated styles for editable-field

The editable-field component will always show the edit-pencil icon but
with a more subdued color and smaller until hover. This allows users to
know that the field is editable without having to hover over all the
known text to figure that out.