Respond: Create/List Remediation Tasks for Incident (#1357)
* Extract notifications and confirmations as mixins

Issue before fix:
* We had built a mechanism for confirmation modals and flash message
notifications into rsa-explorer, but these are now being used elsewhere
in other components. We don’t want to copy and paste this code.

What changed:
* Migrated the confirmation code into a Confirmable mixin, and the
flash message code into a Notifications mixin.

* Add i18n keys for remediation task listing and creation

Added new i18n keys / values to be used in new components:
rsa-remediation-tasks/task, rsa-remediation-tasks/new-task, and
rsa-incident/tasks-and-journal.

* Redux plumbing

* Added a new API call for fetching remediation tasks by incident ID
* Setup new incident creators for managing the new journal/task panel
state, and fetching initialization data (tasks and remediation status
types)
* Added new types
* Updated Incident reducer to respond to remediation related actions

* Minor updates to common components

* The button trigger that we use with power-select in the inspector and
in the new task component needs to have an icon after the label, but
because this is a nested component, we don’t have block content, and so
cannot control the positioning using the block. Updated the trigger
button to have a ‘iconPosition’ property that is either ‘before’ or
‘after’ the label. This allows us to put an icon after the label, which
we need for the new dropdown button look/feel.
* The editable field component can have placeholder text, but I wanted
to be able to style it differently (e.g., with italics). Added the
‘placeholder’ class when the placeholder is showing.

* Task Component

* New Remediation Task component for displaying an individual task with
editable fields.

* New Remediation Task Component

A component representing a new remediation task form that can be saved.
The css for this is in the _task.scss file.

* Incident Container updates

* Updates to incident container component to remove old journal panel
references and to add new state: isShowingTasksAndJournal

* New rsa-incident/tasks-and-journal component

* Represents the panel for showing the journal and the remediation tasks

* Update to Incident Toolbar

* Adds the button for toggling the tasks-and-journal panel

* Removed japanese from dummy env moment config

* Updated pre-existing tests

* Deleted old journal component test, since this will be rewritten
* Updated toolbar test for checking that the toggle action is called
for tasks-and-journal

* Minor fixes

* Removed references to isJournalPanelOpen in the reducer
* Removed action and action creator references to toggling journal panel
* Fixed some english text with i18n translations
* Moved the task and journal element <box> in the container template
and renamed <vbox>
* Updated initializeIncidents reducer to handle view state
* Fixed ui test after removal of isJournalPanelOpen and related actions