Respond: Fixes - Remediation Tasks and Misc. (#1379)
* Update i18n files

Issue before fix:
* We were using a custom moment call that would provide duration that
the remediation task has been open (e.g., with a label that said: Open
for: 12 days). And this was causing some problems with moment, which
would default to japanese translation of the label.

What changed:
* Turns out the rsa-datetime component will do this for us, but would
appear as:  (12 days ago). So switching to use that, instead of a
custom moment computed property. And changed label so it now says:
Opened: (12 days ago).
* Also added some js docs for one of the functions in the task
component.

* Editable Field support for disabled field

Issue before fix:
* The editable field component (which behaves a bit like JIra’s
editable field), did not support disabling the field.

What changed:
* Added an isDisabled property and adjusted the component and template
accordingly. Now, if the field is disabled, the text appears, but no
icon and no handler to swap the text into an input.

* Journal and Tasks Counts / Disable New Task button on closed incident

Issue before fix:
* The Add New Task button for creating a remediation task should be
disabled if the incident is closed.
* We want to show the counts of the journal entries and the remediation
tasks in the tabs of the tasks-and-journal-panel.

What changed:
* Added a couple of computed properties for getting the journal entry
and task counts.
* Updated the template to show the journal entry and task counts in the
tabs.
* Added a computed property for determining whether the incident is
open.
* Updated the template to ensure that if the incident is not open, the
add new task button is disabled.

* Remove Status dropdown from new task form

Issue before fix:
* Turns out that the micro service won’t let the caller set the
remediation task status on creation. Best course of action (confirmed
with PO) is simply to remove the status picker from the new task
creation form.

What changed:
* Removed status picker power select and associated component code from
the new-task component.

* Remapped user disabled property to new name

Issue before fix:
* Whenever a user came back from the service and was displayed in a
dropdown, the dropdown component was disabling the user from selection
if the user object included the disabled: true property. We don’t
always want to do that and want to remove the naming collision there.

What changed:
* Added a remapping function in the reducer to map the disabled
property and value to a new property (isInactive), and deleted the
disabled property so that it cannot trigger UI component handling that
we don’t want.
* Updated test and mock data.

* Removed ID from Alert Inspector metadata

Issue before fix:
* Azka requested that the ID of the alert be removed since it is just a
guid and serves no value to the end user

What changed:
* Removed the display of the alert ID from the inspector

* Disable priority/assignee changes if task is closed

Issue before fix:
* User could change priority and assignee on a remediation task even if
it was closed.

What changed:
* Disabled the priority picker and the assignee editable field when the
task is in a closed state

* Minor CSS updates

* Updated the css for metadata that has disabled state
* Updated the editable field component’s look/feel when disabled
* Adjusted the size of the remediation task header values