Respond: Miscellaneous UI Fixes (#1495)
* Context Hub Tooltip: Decrease font-size on Machine/LiveConnect

Issue before fix:
* In a previous PR we overrode the font-size of the context-hub
tabs/links so that the word UNKNOWN would fit completely on in the tab,
but there is also the possibility that SUSPICIOUS can appear there, and
it is currently truncated (with ellipsis)

What changed:
* Decreased the font-size to 90% to allow for full display of
SUSPICIOUS (the longest term that will appear the context hub tooltip)

Issue: ASOC-38229, ASOC-37150

* ASOC-37981: Navigating to incident can show incorrect details view

Issue before fix:
* If you entered the incident details page with the data sheet events
view active, switched to the nodal graph, and then searched for a
related indicator, and finally clicked on the link to another incident,
you would get reverted back to the data sheet events view instead of
the nodal graph

What changed:
* The problem was with some logic in the INITIALIZE_INCIDENT reducer
which tried to set `hideViz` property from state or from initial state
if it doesn’t exist. The problem is that since `hideVize` is a boolean
and can be (properly) `false`, whenever it was false, the property
would be set from initial state instead, even though we wanted the
false value from current state.
* Updated both `hideViz` and `isShowingTasksAndJournal` properties to
only look at state, since they are booleans and there is no reducer
function that would set them to any other (undefined/null) value.

Issue: ASOC-37981

* ASOC-37994: Cannot resize last column due to column picker

Issue before fix:
* There are times when there are a large number of columns and the
screen is smaller, that the column picker control covers some portion
of the last column. Since this resizer of the column is on the very
right of the cell, this means the column picker completely covers the
resizer, and the user cannot resize

What changed:
* Although I think this issue is more a minor issue with the data grid
component’s implementation, we can provide a fix that (while not
perfect) provides a better experience.
* By setting the resizer’s right position to 30px (the width of the
column picker) on the very last column, we can ensure that the resizer
is always available to the user even on the last column.
* The downside to this approach is that when there is ample space in
the grid, the resizer technically is not at the very right of the
header cell (but instead 30px off). I think this is acceptable because
we don’t actually show the right cell border on the last column cell
anyway, so the user cannot really see that the resizer is not at the
very edge of the cell. Alternative fixes would be more complicated and
are probably not worth the investment (at least on the Respond side). I
would suggest that if we need a more sophisticated fix, this fix occur
in the data-table component itself. That, however, would be a higher
risk change.

* Respond: incidents updateRecord socketRoute incorrect (#1496)

Issue before fix:
* After a PR on the service side got merged, the client could no longer
update incidents.
* Issue was identified as an incorrect socketRoute for
incidents.updateRecord. From what we can tell, the service PR changed
the expected topic id that the client should subscribe to, which was
causing this to break in the client

What changed:
* Updated the subscriptionDestination for incidents.updateRecord to be
the correct value

Issue: ASOC-38231

* enable-column-selector class

Per GB’s recommendation, adding an additional class name to the data
table header when enableColumnSelector is true so that we can more
directly target the resizer fix and even support item sheets without
the column selector enabled.