Tethered Panel: Reset position on display (#1523)
Issue before fix:
* The `position` property on the rsa-content-tethered-panel was being
changed as part of the forceWithinWindow() function. This resulted in
the tethered panel being reopened in the adjusted position every single
subsequent time it was displayed, even when there was no potential for
collision with the window edges in the subsequent display. E.g., in the
case of the context hub tooltip which is activated by many different
triggers on the page, if the first trigger was close to the top of the
window, the tooltip would open (correctly) below the trigger to avoid
collision with the window’s edge. But afterwards, all tooltips would
open on the bottom, even if the trigger was in the middle of the page
and not in danger of collision.

What changed:
* Rather than have forceWithinWindow() change the `position` property
directly, it now changes a private property `_position`. And this
private property is reset from the original `position` property every
time the tethered panel is displayed, so we know it gets reverted to
the originally configured position even if it was adjusted on the last
display.

Issue: ASOC-38715