Respond Filter Checkboxes: Labels extend across panel (#1541)
Issue before fix:
* The labels that wrap all of the checkboxes in the filter panels (in
Incidents, Alerts, and Tasks) were set via css to be display: block.
But this causes a problem that if you click anywhere in the panel on
the same row as a checkbox, it will select it, even if you are clicking
in what appears to be empty space. The label element which is extending
across that empty space handles the click like a selection.
* The Alert Names filter checkboxes in the Alerts page were bouncing
around when you hovered over them due to hover handling in the data
table css. There was also a black background when you hovered on the
row (also from data table css)

What changed:
* Removed the css that set each label to be displayed block in the
filters
* Wrapped all checkboxes in a div, so that they don’t appear on the
same line. This way, the label is inline, but the wrapping div creates
the new line that we’re looking for
* Added some additional css to fix the bouncing of the alert name rows
as you hover in the alert name filters area.

Issue: None (communicated w/ QE)